requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
streamlit==1.28.0
faiss-cpu==1.7.4
tavily-python==0.7.2
python-dotenv==1.0.0
pypdf==3.17.0
langchain-community==0.0.11
pycountry==22.3.5
streamlit-chat==0.1.0